requirements.txt
UnityPy==1.10.0
Pillow==9.5.0
